requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-leaflet==0.26.0
djangorestframework==3.11.0
djangorestframework-gis==0.15
filelock==3.0.12
gunicorn==20.0.4
Markdown==3.2.2
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.15
virtualenvwrapper-win==1.2.6
wfastcgi==3.0.0
whitenoise==5.0.1
